requirements.txt
google-genai==1.63.0
google-adk
mcp
python-dotenv
